exported functions to entry point executable
