--- client-repo/deps.ts
@@ -1,4 +1,0 @@
-export {
-    Application,
-    Router
-} from 'https://deno.land/x/oak@v11.1.0/mod.ts';
