+++ parse/detect_endpoints.ts
